Fix race condition due to closing notifications
Only mark a notification as cleared when the user clicks on it, not when
it is closed automatically after a timeout. This prevents the case where
the user harvests and plants new crops without first clicking the
notification. The new crop timer is started, but then the notification
is closed which itself sets the relevent timer to "Cleared". For crops
this causes the timer to take on its first value but then never
decrease.

Also adds (but commented out) a new debug button to the UI that
dumps the current state of all action controls to the console.