Make queued speech work again
To try and work around the issue where speech gets stuck and never
resets, previously it called synth.cancel() on every speak() call. This
means that if two announcements are made in quick succession, the first
one gets cancelled mid-speech or even before it can start.

This change removes that call to synth.cancel() and instead sets a timer
when speech starts, if it detects that there is already speech in
progress. If the timer expires without the speech finishing it calls
cancel and logs it.